+++ Vic2ToHoI4/Source/HOI4World/HoI4World.h
@@ -135,7 +135,7 @@ class HoI4World
 		void addGreatPowerVPs(const CountryMapping& countryMap);
 		void addStrengthVPs();
 		double getStrongestCountryStrength();
-		double calculateStrengthVPs(HoI4Country* country, double greatestStrength);
+		int calculateStrengthVPs(HoI4Country* country, double greatestStrength);
 
 		vector<int>					getPortProvinces(const vector<int>& locationCandidates);
 		vector<int>					getPortLocationCandidates(const vector<int>& locationCandidates, const HoI4AdjacencyMapping& HoI4AdjacencyMap);
